+++ valle/__init__.py
@@ -1,1 +1,1 @@
-from . import data, modules, utils
+from . import data, models, modules, utils
