--- valle/modules/__init__.py
@@ -1,1 +1,0 @@
-from .model import add_model_arguments, get_model
